Missing joints correlation matrix for JAAD + OpenPose
In the dataset used to make the predictions, in most pedestrian skeletons there are more than 3 undetected joints whose coordinates are set to 0. This leads to lower than desired classification results. Even many skeletons in the dataset have not been detected and the coordinates of all their joints are set to 0. This notebook analyzes which joints are missing and what correlation there is between the joints that are missing at the same time in the same skeleton.