2 autoencoders for skeletons joints recovering
Implemented 2 autoencoders for the recovery of the missing joints in the skeletons of the pedestrians in the dataset. One is a fully-connected autoencoder and the other is a Graph Convolutional Autoencoder.

I have also designed my own targeted denoising function that adds noise most likely to node 8 (the neck) and its nearby nodes. This is so because if the neck is missing, the classification results between crossing and not-crossing are very bad. In this way, when using to predict the resulting autoencoder skeletons with my targeted denoising, the classification results are better than using random denoising.