Pedestrians skeletons missing joints recovering
Recovery of missing joints in skeletons using a denoising autoencoder. The notebook also shows the 2 types of denoising that can be used with my classes: random denoising that has an equal probability of adding noise to each node, or my targeted denoising approach that in this case adds noise more likely to node 8 (the neck) and nearby nodes, since it is the most crucial node when it comes to classifying between crossing and not-crossing, at least in CARLA.

My targeted denoising allows to retrieve skeletons that give better classification results than random denoising.